date이외 컬럼 타입 변경

+++ csv_to_parquet.py
@@ -225,7 +225,18 @@ def store_csv_to_db_with_pgfutter(csv_file, target_table="stock_data"):
 
         cur.execute(f"""
             INSERT INTO {target_table} (date, open, high, low, close, volume, dividends, stock_splits, ticker, capital_gains)
-            SELECT date::DATE, open, high, low, close, volume, dividends, stock_splits, ticker, capital_gains FROM {table_name};
+            SELECT 
+                date::DATE, 
+                open::NUMERIC, 
+                high::NUMERIC, 
+                low::NUMERIC, 
+                close::NUMERIC, 
+                volume::NUMERIC, 
+                dividends::NUMERIC, 
+                stock_splits::NUMERIC, 
+                ticker, 
+                capital_gains::NUMERIC
+            FROM {table_name};
         """)
         conn.commit()
